Action redux unit tests moved; global Api mocks now under beforeEach()
