--- src/redux/redux.spec.js
@@ -1,224 +1,0 @@
-import {
-  ACTIONS,
-  fetchCountryStats,
-  fetchCovidNews,
-  fetchSummary,
-  setCountryTracked,
-} from './actions';
-import configureMockStore from 'redux-mock-store';
-import thunk from 'redux-thunk';
-import {ASYNC_STORAGE_TRACKED_KEY} from '../strings';
-import AsyncStorage from '@react-native-community/async-storage';
-import {CovidApi} from '../api/covid-api';
-import {reducer} from './reducer';
-import {article} from '../screens/news-feed-screen/news-feed-screen.spec';
-import {NewsApi} from '../api/news-api';
-
-describe('Redux unit tests', () => {
-  describe('reducer', () => {
-    it('should return the initial state by default', () => {
-      expect(reducer()).toEqual(initialState);
-    });
-
-    it('sets the historical data with date labels for a chosen country on ACTIONS.GET_COUNTRY_STATS_SUCCESS', () => {
-      const action = {
-        type: ACTIONS.GET_COUNTRY_STATS_SUCCESS,
-        payload: {data, labels},
-      };
-      const expected = {...initialState, labels, data, pending: false};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('sets the error on ACTIONS.GET_COUNTRY_STATS_FAILURE', () => {
-      const action = {type: ACTIONS.GET_COUNTRY_STATS_FAILURE, error};
-      const expected = {...initialState, error, pending: false};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('starts country data fetch with ACTIONS.GET_COUNTRY_STATS_PENDING', () => {
-      const action = {type: ACTIONS.GET_COUNTRY_STATS_PENDING};
-      const expected = {...initialState, pending: true};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('sets fetched articles for related news on ACTIONS.GET_COVID_NEWS_SUCCESS', () => {
-      const action = {
-        type: ACTIONS.GET_COVID_NEWS_SUCCESS,
-        payload: {articles},
-      };
-      const expected = {...initialState, articles, pending: false};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('sets error object after article fetch failure on ACTIONS.GET_COVID_NEWS_FAILURE', () => {
-      const action = {type: ACTIONS.GET_COVID_NEWS_FAILURE, error};
-      const expected = {...initialState, error, pending: false};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('starts fetching news feed articles with ACTIONS.GET_COVID_NEWS_PENDING', () => {
-      const action = {type: ACTIONS.GET_COVID_NEWS_PENDING};
-      const expected = {...initialState, pending: true};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('sets fetched global data on ACTIONS.GET_SUMMARY_SUCCESS', () => {
-      const action = {
-        type: ACTIONS.GET_SUMMARY_SUCCESS,
-        payload: {countries, globalData},
-      };
-      const expected = {...initialState, countries, globalData, pending: false};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('sets error object after failing to fetch global data with ACTIONS.GET_SUMMARY_FAILURE', () => {
-      const action = {type: ACTIONS.GET_SUMMARY_FAILURE, error};
-      const expected = {...initialState, error, pending: false};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    it('starts fetching global data on ACTIONS.GET_SUMMARY_PENDING', () => {
-      const action = {type: ACTIONS.GET_SUMMARY_PENDING};
-      const expected = {...initialState, pending: true};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-
-    describe('sets a country as being (un)tracked with ACTIONS.SET_COUNTRY_TRACKED', () => {
-      it('works correctly with value=true', () => {
-        const actionTrackedTrue = {
-          type: ACTIONS.SET_COUNTRY_TRACKED,
-          payload: {value: true, countrySlug},
-        };
-        const expected = {...initialState, tracked: [countrySlug]};
-
-        expect(reducer(initialState, actionTrackedTrue)).toEqual(expected);
-      });
-
-      it('works correctly with value=false', () => {
-        const actionTrackedFalse = {
-          type: ACTIONS.SET_COUNTRY_TRACKED,
-          payload: {value: false, countrySlug},
-        };
-        const expected = {...initialState};
-
-        expect(
-          reducer(
-            {...initialState, tracked: [countrySlug]},
-            actionTrackedFalse,
-          ),
-        ).toEqual(expected);
-      });
-    });
-
-    it('gets tracked country list from async storage on ACTIONS.GET_ASYNC_STORAGE_TRACKED', () => {
-      const action = {
-        type: ACTIONS.GET_ASYNC_STORAGE_TRACKED,
-        payload: {tracked},
-      };
-      const expected = {...initialState, tracked};
-
-      expect(reducer(undefined, action)).toEqual(expected);
-    });
-  });
-
-  describe('actions', () => {
-    let store;
-
-    beforeEach(() => {
-      store = configureMockStore([thunk])();
-    });
-
-    it('should dispatch expected actions with setCountryTrackedAction', async () => {
-      const expectedActions = [
-        {type: ACTIONS.SET_COUNTRY_TRACKED, payload: {countrySlug, value}},
-      ];
-      await store.dispatch(setCountryTracked(countrySlug, value));
-
-      expect(store.getActions()).toEqual(expectedActions);
-    });
-
-    it('should dispatch expected actions with fetchSummaryAction', async () => {
-      const expectedActions = [
-        {type: ACTIONS.GET_SUMMARY_PENDING},
-        {type: ACTIONS.GET_ASYNC_STORAGE_TRACKED, payload: {tracked}},
-        {type: ACTIONS.GET_SUMMARY_SUCCESS, payload: {countries, globalData}},
-      ];
-
-      await AsyncStorage.setItem(
-        ASYNC_STORAGE_TRACKED_KEY,
-        JSON.stringify(tracked),
-      );
-      await store.dispatch(fetchSummary());
-
-      expect(store.getActions()).toEqual(expectedActions);
-    });
-
-    it('should dispatch expected actions with fetchCountryStatsAction', async () => {
-      const expectedActions = [
-        {type: ACTIONS.GET_COUNTRY_STATS_PENDING},
-        {type: ACTIONS.GET_COUNTRY_STATS_SUCCESS, payload: {data, labels}},
-      ];
-
-      await store.dispatch(fetchCountryStats());
-
-      expect(store.getActions()).toEqual(expectedActions);
-    });
-
-    it('should dispatch expected actions with fetchCovidNewsAction', async () => {
-      const expectedActions = [
-        {type: ACTIONS.GET_COVID_NEWS_PENDING},
-        {type: ACTIONS.GET_COVID_NEWS_SUCCESS, payload: {articles}},
-      ];
-
-      await store.dispatch(fetchCovidNews());
-
-      expect(store.getActions()).toEqual(expectedActions);
-    });
-  });
-});
-
-const countrySlug = 'testSlug';
-const tracked = [countrySlug];
-const value = true;
-const globalData = {
-  NewConfirmed: 123,
-};
-const countries = [
-  {
-    Country: 'Atlantis',
-  },
-];
-const labels = ['06-01', '06-02'];
-const data = [123, 456];
-const articles = [article];
-const initialState = {
-  articles: [],
-  countries: [],
-  tracked: [],
-  globalData: {
-    newConfirmed: 0,
-    newDeaths: 0,
-    newRecovered: 0,
-    totalConfirmed: 0,
-    totalDeaths: 0,
-    totalRecovered: 0,
-  },
-  error: null,
-  labels: [],
-  data: [0],
-  pending: true,
-};
-const error = new Error();
-
-CovidApi.getCountryStats = jest.fn().mockResolvedValue({labels, data});
-CovidApi.getSummary = jest.fn().mockResolvedValue({globalData, countries});
-NewsApi.getTopCovidHeadlines = jest.fn().mockResolvedValue({articles});
